separate LocationInput into its own file
