+++ src/pages/settings/DeliverySettings.js
@@ -13,67 +13,9 @@ import {
   Input,
   Form,
   FormGroup,
-  InputGroup,
-  InputGroupText,
-  InputGroupAddon,
 } from "reactstrap";
 
-function LocationInput(props) {
-  const { fee, location, handleInputChange, index, removeInput } = props;
-
-  return (
-    <>
-      <Col lg="6">
-        <FormGroup>
-          <label className="form-control-label" htmlFor="location">
-            Location
-          </label>
-          <Input
-            className="form-control-alternative"
-            id="location"
-            placeholder="Manila"
-            type="text"
-            value={location}
-            onChange={(e) => handleInputChange(e, index)}
-          />
-        </FormGroup>
-      </Col>
-      <Col
-        lg="6"
-        className="mb-3 mb-lg-0 d-flex align-items-center justify-content-between"
-      >
-        <FormGroup className="w-75">
-          <label className="form-control-label" htmlFor="fee">
-            Fee
-          </label>
-          <InputGroup>
-            <InputGroupAddon addonType="prepend">
-              <InputGroupText>Php</InputGroupText>
-            </InputGroupAddon>
-            <Input
-              className="form-control-alternative"
-              id="fee"
-              type="number"
-              value={fee}
-              onChange={(e) => handleInputChange(e, index)}
-            />
-          </InputGroup>
-        </FormGroup>
-
-        <Button
-          outline
-          className="btn-icon btn-3"
-          color="danger"
-          type="button"
-          style={{ height: "min-content", padding: "2px 4px" }}
-          onClick={() => removeInput(index)}
-        >
-          <i className="ni ni-fat-remove" />
-        </Button>
-      </Col>
-    </>
-  );
-}
+import LocationInput from "./LocationInput";
 
 function DeliverySettings() {
   const [address, setAddress] = useState("");
